+++ src/modules/kiosks/TempLog.js
@@ -1,41 +1,38 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { Grid, Segment } from 'semantic-ui-react';
 
 import Breadcrumbs from 'modules/shared/components/Breadcrumbs';
 import Loader from 'modules/shared/components/Loader';
-import history from 'lib/history';
 import { getKioskSingle } from './selectors';
 import TempLogVisualization from './components/TempLogVisualization';
 import TempLog from './components/TempLog';
+import { getKiosk } from './actions';
 
-const Log = ({ kiosk, isLoading }) => {
-  const links = kiosk
-    ? [
-      {
-        name: 'Home',
-        link: '/dashboard',
-      },
-      {
-        name: 'Kiosks',
-        link: '/kiosks',
-      },
-      {
-        name: kiosk.name,
-        link: `/kiosks/detail/${kiosk._id}`,
-      },
-    ]
-    : [];
-  const backLink = kiosk
-    ? {
-      name: 'Back to kiosk detail',
-      link: `/kiosks/detail/${kiosk._id}`,
-    }
-    : null;
-
-  if (kiosk === null) {
-    history.push('/kiosks');
+const Log = ({ getKiosk, kiosk, isLoading, ...props }) => {
+  const links = [
+    {
+      name: 'Home',
+      link: '/dashboard',
+    },
+    {
+      name: 'Kiosks',
+      link: '/kiosks',
+    },
+    {
+      name: kiosk === null ? '' : kiosk.name,
+      link: `/kiosks/detail/${kiosk === null ? props.match.params.id : kiosk._id}`,
+    },
+  ]
+  const backLink = {
+    name: 'Back to kiosk detail',
+    link: `/kiosks/detail/${kiosk === null ? props.match.params.id : kiosk._id}`,
   }
+  useEffect(() => {
+    if (kiosk === null) {
+      getKiosk(props.match.params.id);
+    }
+  }, []);
 
   return (
     <>
@@ -52,7 +49,7 @@ const Log = ({ kiosk, isLoading }) => {
             </Segment>
           </Grid.Column>
         </Grid.Row>
-        <TempLogVisualization />
+        <TempLogVisualization {...props} />
         <TempLog />
       </Grid>
     </>
@@ -61,7 +58,10 @@ const Log = ({ kiosk, isLoading }) => {
 
 const mapStateToProps = state => ({
   kiosk: getKioskSingle(state),
-  isLoading: state.kiosks.isLoading,
+  isLoading: state.kiosks.isLoading
 });
+const mapDispatchToProps = {
+  getKiosk
+};
 
-export default connect(mapStateToProps)(Log);
+export default connect(mapStateToProps, mapDispatchToProps)(Log);