+++ src/modules/kiosks/components/TempLog.js
@@ -32,7 +32,7 @@ const columns = [
 
 
 const TempLogGrid = ({ temperatureLogs }) => {
-
+  console.log(temperatureLogs)
   return (
     <Grid.Row>
       <Grid.Column>